requirements.txt
blinker==1.8.2
cachelib==0.9.0
circuitbreaker==2.0.0
click==8.1.7
colorama==0.4.6
Deprecated==1.2.14
Flask==3.0.3
Flask-Caching==2.3.0
Flask-Cors==4.0.1
Flask-Limiter==3.8.0
flask-marshmallow==1.2.1
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
importlib_resources==6.4.0
itsdangerous==2.2.0
Jinja2==3.1.4
limits==3.13.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.3
marshmallow-sqlalchemy==1.0.0
mdurl==0.1.2
mysql-connector-python==9.0.0
ordered-set==4.1.0
packaging==24.1
pip==24.0
Pygments==2.18.0
PyJWT==2.9.0
python-dotenv==1.0.1
rich==13.7.1
SQLAlchemy==2.0.31
typing_extensions==4.12.2
Werkzeug==3.0.3
wrapt==1.16.0